# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lead Service",
    "description": "Manages lead buffering, deduplication, cursor pagination, and enrichment for campaign outreach.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3006"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "components": {
    "schemas": {
...
      "BufferNextResponse": {
        "type": "object",
        "properties": {
          "found": {
            "type": "boolean",
            "description": "True when a lead was claimed and returned. False when no lead can be served right now.",
            "example": true
          },
          "lead": {
            "$ref": "#/components/schemas/ServedLead"
          },
          "reason": {
            "type": "string",
            "enum": [
              "credit_insufficient"
            ],
            "description": "Optional reason when found=false. credit_insufficient means the org has no available platform credits, so no paid enrichment/search/LLM action was performed.",
            "example": "credit_insufficient"
          }
        },
        "required": [
          "found"
        ],
        "description": "Response from POST /orgs/buffer/next. When found is true, lead contains the served lead with full canonical FullLead payload under `lead.data`.",
        "examples": [
          {
            "summary": "Apollo lead",
            "value": {
              "found": true,
              "lead": {
                "leadId": "00000000-0000-0000-0000-000000000001",
                "email": "sara@cascobay.com",
                "data": {
                  "leadId": "00000000-0000-0000-0000-000000000001",
                  "apolloPersonId": "5f2a3b4c5d6e7f8a9b0c1d2e",
                  "firstName": "Sara",
                  "lastName": "Freshley",
                  "name": "Sara Freshley",
                  "headline": "Founder at Casco Bay",
                  "linkedinUrl": "https://linkedin.com/in/sara-freshley",
                  "photoUrl": null,
                  "city": "Portland",
                  "state": "ME",
                  "country": "USA",
                  "seniority": "founder",
                  "departments": [
                    "c_suite"
                  ],
                  "subdepartments": null,
                  "functions": null,
                  "twitterUrl": null,
                  "githubUrl": null,
                  "facebookUrl": null,
                  "enrichedAt": "2026-01-01T00:00:00.000Z",
                  "currentTitle": "Founder",
                  "organization": {
                    "id": "10000000-0000-0000-0000-000000000001",
                    "apolloOrganizationId": "5f2a3b4c5d6e7f8a9b0c1d2e",
                    "name": "Casco Bay",
                    "primaryDomain": "cascobay.com",
                    "websiteUrl": "https://cascobay.com",
                    "industry": "marketing",
                    "estimatedNumEmployees": 12,
                    "annualRevenue": "1000000",
                    "logoUrl": null,
                    "shortDescription": null,
                    "linkedinUrl": null,
                    "twitterUrl": null,
                    "facebookUrl": null,
                    "blogUrl": null,
                    "crunchbaseUrl": null,
                    "foundedYear": 2018,
                    "city": "Portland",
                    "state": "ME",
                    "country": "USA",
                    "streetAddress": null,
                    "postalCode": null,
                    "technologyNames": [
                      "GA4"
                    ],
                    "industries": [
                      "marketing"
                    ],
                    "secondaryIndustries": null,
                    "latestFundingStage": "series_a",
                    "latestFundingRoundDate": "2024-06-01",
                    "totalFunding": "5000000",
                    "totalFundingPrinted": "$5M",
                    "fundingEvents": [],
                    "retailLocationCount": null,
                    "publiclyTradedSymbol": null,
                    "publiclyTradedExchange": null,
                    "primaryPhone": null,
                    "seoDescription": null,
                    "angellistUrl": null,
                    "numSuborganizations": null,
                    "alexaRanking": null,
                    "keywords": null
                  },
                  "contacts": [
                    {
                      "channel": "email",
                      "value": "sara@cascobay.com",
                      "status": "verified",
                      "source": "apollo"
                    }
                  ],
                  "employmentHistory": [
                    {
                      "organizationId": "10000000-0000-0000-0000-000000000001",
                      "organizationName": "Casco Bay",
                      "title": "Founder",
                      "startDate": "2018-01-01",
                      "endDate": null,
                      "current": true,
                      "description": null
                    }
                  ]
                },
                "brandIds": [
                  "20000000-0000-0000-0000-000000000001"
                ],
                "orgId": "30000000-0000-0000-0000-000000000001",
                "userId": "40000000-0000-0000-0000-000000000001",
                "apolloPersonId": "5f2a3b4c5d6e7f8a9b0c1d2e"
              }
            }
          },
          {
            "summary": "Buffer exhausted",
            "value": {
              "found": false
            }
          },
          {
            "summary": "Insufficient credits",
            "value": {
              "found": false,
              "reason": "credit_insufficient"
            }
          }
        ]
      },
      "ServedLead": {
        "type": "object",
        "properties": {
          "leadId": {
            "type": "string",
            "format": "uuid",
            "description": "Internal lead UUID. Same as data.leadId — kept at the top level for backwards compatibility with workflow scripts that read it directly.",
            "example": "00000000-0000-0000-0000-000000000001"
          },
          "email": {
            "type": "string",
            "description": "The email address selected for outreach. Always populated when found=true. Same address appears in data.contacts for the 'email' channel.",
            "example": "sara@cascobay.com"
          },
          "data": {
            "$ref": "#/components/schemas/FullLead"
          },
          "brandIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Brand UUIDs this lead was buffered for (echoed back from x-brand-id header).",
            "example": [
              "20000000-0000-0000-0000-000000000001"
            ]
          },
          "orgId": {
            "type": "string",
            "nullable": true,
            "description": "Internal organization UUID owning the campaign.",
            "example": "30000000-0000-0000-0000-000000000001"
          },
          "userId": {
            "type": "string",
            "nullable": true,
            "description": "Internal user UUID who triggered the campaign run.",
            "example": "40000000-0000-0000-0000-000000000001"
          },
          "apolloPersonId": {
            "type": "string",
            "nullable": true,
            "description": "Apollo person ID — same value as data.apolloPersonId.",
            "example": "5f2a3b4c5d6e7f8a9b0c1d2e"
          },
          "goal": {
            "type": "string",
            "nullable": true,
            "description": "Explicit active goal tag for this served lead. null means unattributed.",
            "example": "signup"
          },
          "activeGoalId": {
            "type": "string",
            "nullable": true,
            "description": "Explicit active goal ID tag for this served lead. null means unattributed.",
            "example": "goal_123"
          },
          "brandProfileId": {
            "type": "string",
            "nullable": true,
            "description": "Explicit brand profile ID tag for this served lead. null means unattributed.",
            "example": "brand_profile_123"
          },
          "audienceId": {
            "type": "string",
            "nullable": true,
            "description": "Audience ID (human-service audience.id) this served lead is attributed to. null means unattributed.",
            "example": "audience_123"
          }
        },
        "required": [
          "leadId",
          "email",
          "data",
          "brandIds",
          "orgId",
          "userId"
        ],
        "description": "A single lead served from the campaign buffer. The full lead payload lives under `data` (FullLead shape)."
      },
      "FullLead": {
        "type": "object",
        "properties": {
          "leadId": {
            "type": "string",
            "format": "uuid",
            "description": "Internal lead UUID (lead-service registry). Stable across enrichment refreshes.",
            "example": "00000000-0000-0000-0000-000000000001"
          },
          "apolloPersonId": {
            "type": "string",
            "nullable": true,
            "description": "Apollo person ID — present when the lead was sourced or enriched via Apollo.",
            "example": "5f2a3b4c5d6e7f8a9b0c1d2e"
          },
          "firstName": {
            "type": "string",
            "description": "Lead's first name. Required — lead-service refuses to register a lead without one. Use this for recipientFirstName on outbound email.",
            "example": "Sara"
          },
          "lastName": {
            "type": "string",
            "description": "Lead's last name. Required. Use this for recipientLastName on outbound email.",
            "example": "Freshley"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "Full display name as provided by source (often 'firstName lastName' but not always).",
            "example": "Sara Freshley"
          },
          "headline": {
            "type": "string",
            "nullable": true,
            "description": "Lead's professional headline / current role line (e.g. LinkedIn-style headline).",
            "example": "Founder at Casco Bay"
          },
          "linkedinUrl": {
            "type": "string",
            "nullable": true,
            "description": "Lead's LinkedIn profile URL.",
            "example": "https://linkedin.com/in/sara-freshley"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true,
            "description": "Lead's profile photo URL.",
            "example": "https://media.licdn.com/photo.jpg"
          },
          "city": {
            "type": "string",
            "nullable": true,
            "description": "Lead's city.",
            "example": "Portland"
          },
          "state": {
            "type": "string",
            "nullable": true,
            "description": "Lead's state / province.",
            "example": "ME"
          },
          "country": {
            "type": "string",
            "nullable": true,
            "description": "Lead's country.",
            "example": "USA"
          },
          "seniority": {
            "type": "string",
            "nullable": true,
            "description": "Seniority bucket from enrichment (e.g. 'founder', 'director', 'vp').",
            "example": "founder"
          },
          "departments": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            },
            "description": "Department classifications.",
            "example": [
              "c_suite"
            ]
          },
          "subdepartments": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            },
            "description": "Subdepartment classifications.",
            "example": [
              "founders"
            ]
          },
          "functions": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            },
            "description": "Job function classifications.",
            "example": [
              "entrepreneurship"
            ]
          },
          "twitterUrl": {
            "type": "string",
            "nullable": true,
            "description": "Lead's Twitter / X profile URL.",
            "example": "https://twitter.com/sara"
          },
          "githubUrl": {
            "type": "string",
            "nullable": true,
            "description": "Lead's GitHub profile URL.",
            "example": "https://github.com/sara"
          },
          "facebookUrl": {
            "type": "string",
            "nullable": true,
            "description": "Lead's Facebook profile URL.",
            "example": "https://facebook.com/sara"
          },
          "enrichedAt": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 timestamp of last successful enrichment. null when the lead was registered without enrichment.",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "currentTitle": {
            "type": "string",
            "nullable": true,
            "description": "Lead's current role title — derived from the leads_organizations row where current=true. Mirrors `employmentHistory[].title` for the current entry; surfaced top-level for convenient template binding (e.g. `recipientTitle` on outbound email). null when the lead has no current employment row or the row has no title.",
            "example": "Founder"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationView"
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactMethodView"
            },
            "description": "All contact methods attached to this lead — email, phone, etc. May be empty.",
            "example": [
              {
                "channel": "email",
                "value": "sara@cascobay.com",
                "status": "verified",
                "source": "apollo"
              }
            ]
          },
          "employmentHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmploymentEntryView"
            },
            "description": "Full employment history (current + past). Returned in insertion order; check the `current` flag to find the present role.",
            "example": [
              {
                "organizationId": "10000000-0000-0000-0000-000000000001",
                "organizationName": "Casco Bay",
                "title": "Founder",
                "startDate": "2018-01-01",
                "endDate": null,
                "current": true,
                "description": null
              }
            ]
          }
        },
        "required": [
          "leadId",
          "apolloPersonId",
          "firstName",
          "lastName",
          "name",
          "headline",
          "linkedinUrl",
          "photoUrl",
          "city",
          "state",
          "country",
          "seniority",
          "departments",
          "subdepartments",
          "functions",
          "twitterUrl",
          "githubUrl",
          "facebookUrl",
          "enrichedAt",
          "currentTitle",
          "organization",
          "contacts",
          "employmentHistory"
        ],
        "description": "Canonical lead representation returned by every lead-bearing endpoint. Built entirely from structured columns — there is no `metadata` or `raw` Apollo passthrough. Field names are stable regardless of upstream enrichment provider.",
        "example": {
          "leadId": "00000000-0000-0000-0000-000000000001",
          "apolloPersonId": "5f2a3b4c5d6e7f8a9b0c1d2e",
          "firstName": "Sara",
          "lastName": "Freshley",
          "name": "Sara Freshley",
          "headline": "Founder at Casco Bay",
          "linkedinUrl": "https://linkedin.com/in/sara-freshley",
          "photoUrl": null,
          "city": "Portland",
          "state": "ME",
          "country": "USA",
          "seniority": "founder",
          "departments": [
            "c_suite"
          ],
          "subdepartments": [
            "founders"
          ],
          "functions": [
            "entrepreneurship"
          ],
          "twitterUrl": null,
          "githubUrl": null,
          "facebookUrl": null,
          "enrichedAt": "2026-01-01T00:00:00.000Z",
          "currentTitle": "Founder",
          "organization": {
            "id": "10000000-0000-0000-0000-000000000001",
            "apolloOrganizationId": "5f2a3b4c5d6e7f8a9b0c1d2e",
            "name": "Casco Bay",
            "primaryDomain": "cascobay.com",
            "websiteUrl": "https://cascobay.com",
            "industry": "marketing",
            "estimatedNumEmployees": 12,
            "annualRevenue": "1000000",
            "logoUrl": "https://logo.clearbit.com/cascobay.com",
            "shortDescription": "Boutique digital marketing agency in Portland, ME.",
            "linkedinUrl": "https://linkedin.com/company/cascobay",
            "twitterUrl": null,
            "facebookUrl": null,
            "blogUrl": null,
            "crunchbaseUrl": null,
            "foundedYear": 2018,
            "city": "Portland",
            "state": "ME",
            "country": "USA",
            "streetAddress": null,
            "postalCode": "04101",
            "technologyNames": [
              "GA4",
              "HubSpot"
            ],
            "industries": [
              "marketing",
              "advertising"
            ],
            "secondaryIndustries": null,
            "latestFundingStage": "series_a",
            "latestFundingRoundDate": "2024-06-01",
            "totalFunding": "5000000",
            "totalFundingPrinted": "$5M",
            "fundingEvents": [
              {
                "id": "fund_5f2a3b4c5d6e7f8a9b0c1d2e",
                "date": "2024-06-01",
                "type": "Series A",
                "investors": "Acme VC, Foo Capital",
                "amount": 5000000,
                "currency": "USD",
                "newsUrl": "https://techcrunch.com/2024/06/01/casco-bay-series-a"
              }
            ],
            "retailLocationCount": null,
            "publiclyTradedSymbol": null,
            "publiclyTradedExchange": null,
            "primaryPhone": "+15555550100",
            "seoDescription": "Casco Bay is a boutique digital marketing agency based in Portland, Maine.",
            "angellistUrl": null,
            "numSuborganizations": 0,
            "alexaRanking": 250000,
            "keywords": [
              "marketing",
              "branding",
              "digital"
            ]
          },
          "contacts": [
            {
              "channel": "email",
              "value": "sara@cascobay.com",
              "status": "verified",
              "source": "apollo"
            }
          ],
          "employmentHistory": [
            {
              "organizationId": "10000000-0000-0000-0000-000000000001",
              "organizationName": "Casco Bay",
              "title": "Founder",
              "startDate": "2018-01-01",
              "endDate": null,
              "current": true,
              "description": null
            }
          ]
        }
      },
...
      "ContactMethodView": {
        "type": "object",
        "properties": {
          "channel": {
            "type": "string",
            "description": "Contact channel kind. Currently used: 'email', 'phone'. Stable identifier — case-sensitive.",
            "example": "email"
          },
          "value": {
            "type": "string",
            "description": "Contact value (the actual email address, phone number, etc.). Unique per (leadId, channel).",
            "example": "sara@cascobay.com"
          },
          "status": {
            "type": "string",
            "nullable": true,
            "description": "Provider-reported status of the contact value (e.g. 'verified', 'unverified', 'extrapolated' for emails). null when not classified.",
            "example": "verified"
          },
          "source": {
            "type": "string",
            "description": "Where this contact method originated (e.g. 'apollo', 'manual', 'csv-upload').",
            "example": "apollo"
          }
        },
        "required": [
          "channel",
          "value",
          "status",
          "source"
        ],
        "description": "One contact endpoint attached to a lead — email, phone, or any other channel. Multiple rows per lead are possible.",
        "example": {
          "channel": "email",
          "value": "sara@cascobay.com",
          "status": "verified",
          "source": "apollo"
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "BufferNextRequest": {
        "type": "object",
        "properties": {},
        "description": "Empty body. The brand, feature, goal, and run identity are read from headers; lead-service resolves the audience (features-service) and serves the next person (human-service). No filters and no provider are accepted — human-service owns both."
      },
...
    },
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authenticating requests. Must match the LEAD_SERVICE_API_KEY env var on the server."
      }
    }
  },
  "paths": {
...
    "/orgs/buffer/next": {
      "post": {
        "summary": "Pull the next lead from the buffer",
        "description": "Claims and returns the next available lead from the campaign buffer. Response contains the full canonical lead payload (FullLead) under `lead.data` — use `data.firstName`, `data.lastName`, `data.organization.name` for outbound recipient fields.",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "API key for authenticating requests"
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Internal organization UUID from client-service"
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Internal user UUID from client-service"
          },
          {
            "in": "header",
            "name": "x-run-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The caller's run ID (used as parentRunId when creating this service's own run)"
          },
          {
            "in": "header",
            "name": "x-campaign-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Campaign identifier (auto-injected by workflow-service)"
          },
          {
            "in": "header",
            "name": "x-brand-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Brand identifier(s), comma-separated for multi-brand campaigns (auto-injected by workflow-service). Example: uuid1,uuid2,uuid3"
          },
          {
            "in": "header",
            "name": "x-workflow-slug",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Workflow slug (auto-injected by workflow-service)"
          },
          {
            "in": "header",
            "name": "x-feature-slug",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Feature slug for tracking (propagated through the call chain)"
          },
          {
            "in": "header",
            "name": "x-goal",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Active goal enum/name for the campaign activity, when explicitly tagged by the caller."
          },
          {
            "in": "header",
            "name": "x-active-goal-id",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Active goal identifier for the campaign activity, when explicitly tagged by the caller."
          },
          {
            "in": "header",
            "name": "x-brand-profile-id",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Brand profile identifier for persona-scoped attribution, when explicitly tagged by the caller."
          },
          {
            "in": "header",
            "name": "x-audience-id",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Audience identifier (human-service audience.id) for attribution, when explicitly tagged by the caller."
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BufferNextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Next lead from buffer (or found=false when exhausted)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BufferNextResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}